% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonClassifyByProfile}
\alias{PRSMonClassifyByProfile}
\title{Retrieve profile classifications}
\usage{
PRSMonClassifyByProfile(
  prsObj,
  tagType,
  profile,
  additionalURLParameters = NULL,
  parser = "json",
  GET.with = "classifybyprofile"
)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{tagType}{Tag type for which the classification is done.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector with named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{additionalURLParameters}{A character vector with named elements each corresponding to
an additional parameter to be passed through the constructed query URL.}

\item{parser}{A string identifying parser to be applied or a parser function.
If NULL no parser is applied.}

\item{GET.with}{A string for the GET command.}

\item{nrecs}{Number of recommendations to be returned.}
}
\value{
A PRSMon object
}
\description{
Classify tag type tags based on profiles of tags.
}
\details{
The result is assigned to \code{prsObj$Value}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{PRSMonRecommendByProfile}()},
\code{\link{PRSMonRecommend}()}
}
\concept{Recommendations computation functions}
