% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonEcho}
\alias{PRSMonEcho}
\title{Echo.}
\usage{
PRSMonEcho(prsObj, x)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{x}{Object to be echoed.}
}
\value{
An PRSMon object
}
\description{
Echoes given argument.
}
\concept{Echo functions}
