% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonEchoFunctionValue}
\alias{PRSMonEchoFunctionValue}
\title{Echo function application to monad's value.}
\usage{
PRSMonEchoFunctionValue(prsObj, f)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{f}{A function to be applied to \code{prsObj$Value}.}
}
\value{
A PRSMon object.
}
\description{
Applies a function to the "Value" element/member of the monad object
and prints the result.
}
\details{
Prints \code{f(prsObj$Value)}.
}
