% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonEchoValue}
\alias{PRSMonEchoValue}
\title{Echo monad's value.}
\usage{
PRSMonEchoValue(prsObj)
}
\arguments{
\item{prsObj}{An PRSMon object.}
}
\value{
A PRSMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{prsObj$Value}.
}
