% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonFailure}
\alias{PRSMonFailure}
\title{Failure producing function}
\usage{
PRSMonFailure(prsObj, message = NULL)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{message}{A message to echoed. If NULL no message is echoed.}
}
\value{
A PRSMon object.
}
\description{
Puts the monad object into a monadic failure.
}
