% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\docType{data}
\name{PRSMonFailureSymbol}
\alias{PRSMonFailureSymbol}
\title{Failure symbol for PRSMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
PRSMonFailureSymbol
}
\description{
Failure symbol for the monad PRSMon.
}
\keyword{datasets}
