% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonGetHistoryQueryString}
\alias{PRSMonGetHistoryQueryString}
\title{From a history specification into a query string}
\usage{
PRSMonGetHistoryQueryString(
  prsObj,
  history,
  functionName = "PRSMonGetHistoryDataFrame",
  warningQ = TRUE
)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{history}{History specification.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{functionName}{A string that is a name of this function or a delegating function.}

\item{warningQ}{Should a warning be issued if \code{history} is of unknown type?}
}
\value{
A PRSMon object
}
\description{
Transforms a history specification into a data frame.
}
\details{
The result data frame is with columns "Score", \code{smr$ItemColumnName};
assigned to \code{prsObj$Value}.
If \code{history = NULL} then \code{prsObj$Value} is used.
}
