% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonGetProfileQueryString}
\alias{PRSMonGetProfileQueryString}
\title{From a profile specification into query string}
\usage{
PRSMonGetProfileQueryString(
  prsObj,
  profile,
  functionName = "PRSMonGetProfileDataFrame",
  warningQ = TRUE
)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{functionName}{A string that is a name of this function or a delegating function.}

\item{warningQ}{Should a warning be issued if \code{profile} is of unknown type?}
}
\value{
A PRSMon object
}
\description{
Transforms a profile specification into a data frame.
}
\details{
The result is assigned to \code{prsObj$Value}.
If \code{profile = NULL} then \code{prsObj$Value} is used.
}
