% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonOption}
\alias{PRSMonOption}
\title{Optional function application to monad's object.}
\usage{
PRSMonOption(prsObj, f)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A PRSMon object.
}
\description{
If monadic failure is obtained from \code{f(prsObj)}
then returns the original \code{prsObj};
else returns the result of \code{f(prsObj)}.
}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
