% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonRecommendByProfile}
\alias{PRSMonRecommendByProfile}
\title{Retrieve profile recommendations}
\usage{
PRSMonRecommendByProfile(
  prsObj,
  profile,
  nrecs = 12,
  additionalURLParameters = NULL,
  parser = "json",
  GET.with = "recommendbyprofile"
)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector with named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{nrecs}{Number of recommendations to be returned.}

\item{additionalURLParameters}{A character vector with named elements each corresponding to
an additional parameter to be passed through the constructed query URL.}

\item{parser}{A string identifying parser to be applied or a parser function.
If NULL no parser is applied.}

\item{GET.with}{A string for the GET command.}
}
\value{
A PRSMon object
}
\description{
Recommend items based on profiles of tags.
}
\details{
The result is assigned to \code{prsObj$Value}.
This function calls \code{\link{RCurl::getURL}}.
The query and the result from that call are stored in
\code{prsObj$LastQuery} and \code{prsObj$LastResult} respectively.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{PRSMonClassifyByProfile}()},
\code{\link{PRSMonRecommend}()}
}
\concept{Recommendations computation functions}
