% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonSetLastQuery}
\alias{PRSMonSetLastQuery}
\title{Set LastQuery.}
\usage{
PRSMonSetLastQuery(prsObj, LastQuery)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{LastQuery}{An object member to be set.}
}
\value{
An PRSMon object.
}
\description{
Sets LastQuery into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{PRSMonSetDatabaseURL}()},
\code{\link{PRSMonSetItemColumnName}()},
\code{\link{PRSMonSetLastResult}()},
\code{\link{PRSMonSetParser}()},
\code{\link{PRSMonSetServerSideObjectID}()},
\code{\link{PRSMonSetServerURL}()},
\code{\link{PRSMonSetValue}()},
\code{\link{PRSMonTakeDatabaseURL}()},
\code{\link{PRSMonTakeItemColumnName}()},
\code{\link{PRSMonTakeLastQuery}()},
\code{\link{PRSMonTakeLastResult}()},
\code{\link{PRSMonTakeParser}()},
\code{\link{PRSMonTakeServerSideObjectID}()},
\code{\link{PRSMonTakeServerURL}()},
\code{\link{PRSMonTakeValue}()}
}
\concept{Set/Take functions}
