% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonSetLastResult}
\alias{PRSMonSetLastResult}
\title{Set LastResult.}
\usage{
PRSMonSetLastResult(prsObj, LastResult)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{LastResult}{An object member to be set.}
}
\value{
An PRSMon object.
}
\description{
Sets LastResult into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{PRSMonSetDatabaseURL}()},
\code{\link{PRSMonSetItemColumnName}()},
\code{\link{PRSMonSetLastQuery}()},
\code{\link{PRSMonSetParser}()},
\code{\link{PRSMonSetServerSideObjectID}()},
\code{\link{PRSMonSetServerURL}()},
\code{\link{PRSMonSetValue}()},
\code{\link{PRSMonTakeDatabaseURL}()},
\code{\link{PRSMonTakeItemColumnName}()},
\code{\link{PRSMonTakeLastQuery}()},
\code{\link{PRSMonTakeLastResult}()},
\code{\link{PRSMonTakeParser}()},
\code{\link{PRSMonTakeServerSideObjectID}()},
\code{\link{PRSMonTakeServerURL}()},
\code{\link{PRSMonTakeValue}()}
}
\concept{Set/Take functions}
