% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonSetServerURL}
\alias{PRSMonSetServerURL}
\title{Set ServerURL.}
\usage{
PRSMonSetServerURL(prsObj, ServerURL)
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{ServerURL}{An object member to be set.}
}
\value{
An PRSMon object.
}
\description{
Sets ServerURL into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{PRSMonSetDatabaseURL}()},
\code{\link{PRSMonSetItemColumnName}()},
\code{\link{PRSMonSetLastQuery}()},
\code{\link{PRSMonSetLastResult}()},
\code{\link{PRSMonSetParser}()},
\code{\link{PRSMonSetServerSideObjectID}()},
\code{\link{PRSMonSetValue}()},
\code{\link{PRSMonTakeDatabaseURL}()},
\code{\link{PRSMonTakeItemColumnName}()},
\code{\link{PRSMonTakeLastQuery}()},
\code{\link{PRSMonTakeLastResult}()},
\code{\link{PRSMonTakeParser}()},
\code{\link{PRSMonTakeServerSideObjectID}()},
\code{\link{PRSMonTakeServerURL}()},
\code{\link{PRSMonTakeValue}()}
}
\concept{Set/Take functions}
