% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonTakeLastQuery}
\alias{PRSMonTakeLastQuery}
\title{Take LastQuery.}
\usage{
PRSMonTakeLastQuery(prsObj, functionName = "PRSMonTakeLastQuery")
}
\arguments{
\item{prsObj}{An PRSMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{PRSMonFailureSymbol}.
}
\description{
Takes LastQuery from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{PRSMonSetDatabaseURL}()},
\code{\link{PRSMonSetItemColumnName}()},
\code{\link{PRSMonSetLastQuery}()},
\code{\link{PRSMonSetLastResult}()},
\code{\link{PRSMonSetParser}()},
\code{\link{PRSMonSetServerSideObjectID}()},
\code{\link{PRSMonSetServerURL}()},
\code{\link{PRSMonSetValue}()},
\code{\link{PRSMonTakeDatabaseURL}()},
\code{\link{PRSMonTakeItemColumnName}()},
\code{\link{PRSMonTakeLastResult}()},
\code{\link{PRSMonTakeParser}()},
\code{\link{PRSMonTakeServerSideObjectID}()},
\code{\link{PRSMonTakeServerURL}()},
\code{\link{PRSMonTakeValue}()}
}
\concept{Set/Take functions}
