% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSMon.R
\name{PRSMonUnit}
\alias{PRSMonUnit}
\title{Make a PRSMon Unit}
\usage{
PRSMonUnit()
}
\value{
An S3 class "PRSMon". In other words, a list with the attribute "class" set to "PRSMon".
}
\description{
Creates a monad object.
}
